requirements.txt
fastapi==0.70.0
pytest==7.0.1
requests==2.27.1
uvicorn==0.15.0
slackclient==2.9.3